Lots of changes and additions
Add lots of documentation.
Remove unnecessary documentation.
Add more operators to Atomic.
Add member function call to Atomic.
Increase Clock tick counter to 64 bits.
Move interrupt vectors into respective files, removing them from user code.
Add version of reverse which has both source and destination arrays.
Make swap() take pointer instead of reference.
Make quicksort a generic helper function.
Remove requirement to pass container size with container. (Array, Queue, etc...)
Remove atomic versions of functions from containers.
Add const to container functions where applicable.
Simplify layout of chip.hpp.
Increase max ns delay to 10ms.
Make Serial use 2x multiplier.
Remove size_t and ssize_t.
Rename optional to Optional.
Make optional contents constant.
Add some tests for quicksort, priority queue, and random.
Add Vector container.

+++ src/i2c.hpp
@@ -1,7 +1,7 @@
+/// # I2C
+
 #ifndef NBOS_I2C_HPP
 #define NBOS_I2C_HPP
-
-/// # I2C
 
 #include <queue.hpp>
 
@@ -69,7 +69,7 @@ struct I2c {
 	static void twiCallback(void* data) {
 		auto queue = (Queue<Action, S>*)data;
 
-		static size_t i;
+		static int i;
 
 		Action action;
